+++ HISTORY.md
@@ -27,6 +27,10 @@
 * **Translation Updates:**
     * German
 
+### 2025-12-17
+* microworlds (EDC) extension: added automatic vertical scrolling to fancy speech & thought bubbles
+* microworlds (EDC) extension: added optional "max height" input to "fancy say / think"  extensions
+
 ### 2025-12-13
 * frequency distribution analysis library: fixed "plot bars" block to handle zero values gracefully